+++ src/apps/update/app.tsx
@@ -1,3 +1,4 @@
+import { useDarkMode } from '../shared/styles';
 import { UpdateModal } from './update';
 import { getCurrentWebviewWindow } from '@tauri-apps/api/webviewWindow';
 import { check, Update } from '@tauri-apps/plugin-updater';
@@ -6,6 +7,8 @@ import { useEffect, useState } from 'react';
 
 export function App() {
   const [update, setUpdate] = useState<Update | null>(null);
+
+  const isDarkMode = useDarkMode();
 
   useEffect(() => {
     const webview = getCurrentWebviewWindow();
@@ -28,13 +31,13 @@ export function App() {
     return null;
   }
 
-  return <ConfigProvider
-    theme={{
-      algorithm: window.matchMedia('(prefers-color-scheme: dark)').matches
-        ? theme.darkAlgorithm
-        : theme.defaultAlgorithm,
-    }}
-  >
-    <UpdateModal update={update}/>
-  </ConfigProvider>;
-};
+  return (
+    <ConfigProvider
+      theme={{
+        algorithm: isDarkMode ? theme.darkAlgorithm : theme.defaultAlgorithm,
+      }}
+    >
+      <UpdateModal update={update} />
+    </ConfigProvider>
+  );
+}